backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==0.3.23
click==8.3.1
clickhouse-connect==0.10.0
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
duckdb==1.4.3
ecdsa==0.19.1
fastapi==0.125.0
filelock==3.20.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
hnswlib==0.8.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.10
llama-index-cli==0.5.3
llama-index-core==0.14.10
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.12
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.5
llama-parse==0.6.54
lz4==4.4.5
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
openai==2.13.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.1
posthog==7.4.0
propcache==0.4.1
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zstandard==0.25.0
razorpay
google-generativeai
langchain-text-splitters
